fix(GeneralForm): Fix pattern validations
In the 'pattern' property, the backslash '\' must be escaped with
another backslash '\\' to be interpreted correctly in the HTML regex.
ESLint warns about this with a message like "Unnecessary escape
character".

For example, [\+] in the prop is passed to the DOM as [+], which throws
an error (special chars like '+' must be escaped because HTML regex is
compiled with the 'v' flag)